+++ src/components/Masonry/index.ts
@@ -1,0 +1,1 @@
+export { default as OrigamMasonry } from './OrigamMasonry.vue'