+++ src/interfaces/index.ts
@@ -171,6 +171,8 @@ export * from './Grids/spacer.interface'
 export * from './Grid/grid.interface'
 export * from './Grid/grid-item.interface'
 
+export * from './Masonry/masonry.interface'
+
 export * from './Icon/icon.interface'
 
 export * from './Img/img.interface'