+++ src/types/index.ts
@@ -139,6 +139,8 @@ export * from './Grids/spacer.type'
 export * from './Grid/grid-flow.type'
 export * from './Grid/grid-align.type'
 
+export * from './Masonry/masonry-align.type'
+
 export * from './Icon/class-icon.type'
 export * from './Icon/component-icon.type'
 export * from './Icon/icon.type'